requirements.txt
annotated-types==0.6.0
anyio==4.2.0
aiomysql==0.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==7.1.2
exceptiongroup==1.2.0
h11==0.14.0
httptools==0.6.1
idna==3.6
importlib-metadata==6.7.0
numpy==1.26.1
packaging==23.2
pandas==2.1.0
pydantic==2.5.3
pydantic_core==2.14.6
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
starlette==0.35.1
typing_extensions==4.9.0
urllib3==2.1.0
fastapi==0.109.0
gunicorn==21.2.0
uvicorn==0.26.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
psutil==5.9.7
httpx==0.26.0